remote/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
decorator==4.4.2
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
incremental==24.7.2
moviepy==1.0.3
msgpack==1.0.8
numpy==2.1.1
pillow==10.4.0
proglog==0.1.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==75.1.0
sqlparse==0.5.1
tomli==2.0.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.0.1
